+++ src/compile/defaultCompiler.ts
@@ -10,19 +10,28 @@ import {
 	createSvelteCompiler,
 	InitialOptions as SvelteCompilerInitialOptions,
 } from './svelteCompiler.js';
+import {
+	createExternalsCompiler,
+	InitialOptions as ExternalsCompilerInitialOptions,
+} from './externalsCompiler.js';
 
 export const createDefaultCompiler = (
 	swcCompilerOptions?: SwcCompilerInitialOptions,
 	svelteCompilerOptions?: SvelteCompilerInitialOptions,
+	externalsCompilerOptions?: ExternalsCompilerInitialOptions,
 	compilerOptions?: CompilerInitialOptions,
 ): Compiler => {
 	const swcCompiler = createSwcCompiler(swcCompilerOptions);
 	const svelteCompiler = createSvelteCompiler(svelteCompilerOptions);
+	const externalsCompiler = createExternalsCompiler(externalsCompilerOptions);
 
 	if (!compilerOptions?.getCompiler) {
 		compilerOptions = {
 			...compilerOptions,
 			getCompiler: (source: CompilationSource) => {
+				if (source.sourceType === 'externals') {
+					return externalsCompiler;
+				}
 				switch (source.extension) {
 					case TS_EXTENSION:
 						return swcCompiler;