+++ src/frontend/bar.ts
@@ -1,0 +1,3 @@
+import {backOut} from 'svelte/easing';
+
+export const bar = backOut(0.5);