+++ src/frontend/foo.ts
@@ -1,0 +1,1 @@
+console.log('src/frontend/foo.ts says hi with import.meta.url:', import.meta.url);